+++ src/main/java/com/newdeal/ledger/account/dto/AccountType.java
@@ -1,0 +1,5 @@
+package com.newdeal.ledger.account.dto;
+
+public enum AccountType {
+	MONEY, ACCOUNT
+}
