+++ src/main/java/com/newdeal/ledger/account/mapper/AccountMapper.java
@@ -1,20 +1,19 @@
-package com.newdeal.ledger.cardaccount.mapper;
+package com.newdeal.ledger.account.mapper;
 
 import java.util.List;
 
 import org.apache.ibatis.annotations.Mapper;
 import org.apache.ibatis.annotations.Param;
 
-import com.newdeal.ledger.cardaccount.dto.AccountDto;
-import com.newdeal.ledger.cardaccount.dto.AccountRequest;
+import com.newdeal.ledger.account.dto.AccountDto;
+import com.newdeal.ledger.account.dto.AccountRequest;
 
 @Mapper
 public interface AccountMapper {
-	List<AccountDto> findAllByEmail(String email);
-
 	void createAccount(@Param("email") String email, @Param("request") AccountRequest.Create request);
 
 	void updateAccount(@Param("accountId") Integer accountId, @Param("request") AccountRequest.Update request);
 
 	void deleteAccount(Integer accountId);
+	List<AccountDto> findAccountsByEmail(String email);
 }
