--- src/main/java/com/newdeal/ledger/cardaccount/service/AccountService.java
@@ -1,17 +1,0 @@
-package com.newdeal.ledger.cardaccount.service;
-
-import java.util.List;
-
-import com.newdeal.ledger.cardaccount.dto.AccountDto;
-import com.newdeal.ledger.cardaccount.dto.AccountRequest;
-
-public interface AccountService {
-
-	List<AccountDto> selectAllByEmail(String email);
-
-	void createAccount(String tempEmail, AccountRequest.Create request);
-
-	void updateAccount(Integer accountId, AccountRequest.Update request);
-
-	void deleteAccount(Integer accountId);
-}
